Separate `ResourceFormat*` methods from main classes
