+++ core/crypto/crypto.h
@@ -32,8 +32,6 @@
 
 #include "core/crypto/hashing_context.h"
 #include "core/io/resource.h"
-#include "core/io/resource_loader.h"
-#include "core/io/resource_saver.h"
 #include "core/object/ref_counted.h"
 
 class CryptoKey : public Resource {
@@ -143,27 +141,4 @@ public:
 	// Compares two PackedByteArrays for equality without leaking timing information in order to prevent timing attacks.
 	// @see: https://paragonie.com/blog/2015/11/preventing-timing-attacks-on-string-comparison-with-double-hmac-strategy
 	bool constant_time_compare(const PackedByteArray &p_trusted, const PackedByteArray &p_received);
-};
-
-class ResourceFormatLoaderCrypto : public ResourceFormatLoader {
-	GDSOFTCLASS(ResourceFormatLoaderCrypto, ResourceFormatLoader);
-
-public:
-	virtual Ref<Resource> load(const String &p_path, const String &p_original_path = "", Error *r_error = nullptr, bool p_use_sub_threads = false, float *r_progress = nullptr, CacheMode p_cache_mode = CACHE_MODE_REUSE) override;
-	virtual void get_recognized_extensions(List<String> *p_extensions) const override;
-	virtual bool handles_type(const String &p_type) const override;
-	virtual String get_resource_type(const String &p_path) const override;
-
-	// Treat certificates as text files, do not generate a `*.{crt,key,pub}.uid` file.
-	virtual ResourceUID::ID get_resource_uid(const String &p_path) const override { return ResourceUID::INVALID_ID; }
-	virtual bool has_custom_uid_support() const override { return true; }
-};
-
-class ResourceFormatSaverCrypto : public ResourceFormatSaver {
-	GDSOFTCLASS(ResourceFormatSaverCrypto, ResourceFormatSaver);
-
-public:
-	virtual Error save(const Ref<Resource> &p_resource, const String &p_path, uint32_t p_flags = 0) override;
-	virtual void get_recognized_extensions(const Ref<Resource> &p_resource, List<String> *p_extensions) const override;
-	virtual bool recognize(const Ref<Resource> &p_resource) const override;
 };
