+++ core/extension/gdextension.h
@@ -32,8 +32,7 @@
 
 #include "core/extension/gdextension_interface.gen.h"
 #include "core/extension/gdextension_loader.h"
-#include "core/io/resource_loader.h"
-#include "core/object/ref_counted.h"
+#include "core/io/resource.h"
 
 class GDExtensionMethodBind;
 
@@ -180,21 +179,6 @@ public:
 };
 
 VARIANT_ENUM_CAST(GDExtension::InitializationLevel)
-
-class GDExtensionResourceLoader : public ResourceFormatLoader {
-	GDSOFTCLASS(GDExtensionResourceLoader, ResourceFormatLoader);
-
-public:
-	static Error load_gdextension_resource(const String &p_path, Ref<GDExtension> &p_extension);
-
-	virtual Ref<Resource> load(const String &p_path, const String &p_original_path, Error *r_error, bool p_use_sub_threads = false, float *r_progress = nullptr, CacheMode p_cache_mode = CACHE_MODE_REUSE) override;
-	virtual void get_recognized_extensions(List<String> *p_extensions) const override;
-	virtual bool handles_type(const String &p_type) const override;
-	virtual String get_resource_type(const String &p_path) const override;
-#ifdef TOOLS_ENABLED
-	virtual void get_classes_used(const String &p_path, HashSet<StringName> *r_classes) override;
-#endif // TOOLS_ENABLED
-};
 
 #ifdef TOOLS_ENABLED
 class GDExtensionEditorPlugins {
