+++ core/io/image_loader.cpp
@@ -139,74 +139,3 @@ void ImageLoader::cleanup() {
 		remove_image_format_loader(loader[0]);
 	}
 }
-
-/////////////////
-
-Ref<Resource> ResourceFormatLoaderImage::load(const String &p_path, const String &p_original_path, Error *r_error, bool p_use_sub_threads, float *r_progress, CacheMode p_cache_mode) {
-	Ref<FileAccess> f = FileAccess::open(p_path, FileAccess::READ);
-	if (f.is_null()) {
-		if (r_error) {
-			*r_error = ERR_CANT_OPEN;
-		}
-		return Ref<Resource>();
-	}
-
-	uint8_t header[4] = { 0, 0, 0, 0 };
-	f->get_buffer(header, 4);
-
-	bool unrecognized = header[0] != 'G' || header[1] != 'D' || header[2] != 'I' || header[3] != 'M';
-	if (unrecognized) {
-		if (r_error) {
-			*r_error = ERR_FILE_UNRECOGNIZED;
-		}
-		ERR_FAIL_V(Ref<Resource>());
-	}
-
-	String extension = f->get_pascal_string();
-
-	int idx = -1;
-
-	for (int i = 0; i < ImageLoader::loader.size(); i++) {
-		if (ImageLoader::loader[i]->recognize(extension)) {
-			idx = i;
-			break;
-		}
-	}
-
-	if (idx == -1) {
-		if (r_error) {
-			*r_error = ERR_FILE_UNRECOGNIZED;
-		}
-		ERR_FAIL_V(Ref<Resource>());
-	}
-
-	Ref<Image> image;
-	image.instantiate();
-
-	Error err = ImageLoader::loader.write[idx]->load_image(image, f);
-
-	if (err != OK) {
-		if (r_error) {
-			*r_error = err;
-		}
-		return Ref<Resource>();
-	}
-
-	if (r_error) {
-		*r_error = OK;
-	}
-
-	return image;
-}
-
-void ResourceFormatLoaderImage::get_recognized_extensions(List<String> *p_extensions) const {
-	p_extensions->push_back("image");
-}
-
-bool ResourceFormatLoaderImage::handles_type(const String &p_type) const {
-	return p_type == "Image";
-}
-
-String ResourceFormatLoaderImage::get_resource_type(const String &p_path) const {
-	return p_path.get_extension().to_lower() == "image" ? "Image" : String();
-}
