+++ core/io/json.h
@@ -31,8 +31,6 @@
 #pragma once
 
 #include "core/io/resource.h"
-#include "core/io/resource_loader.h"
-#include "core/io/resource_saver.h"
 #include "core/variant/variant.h"
 
 class JSON : public Resource {
@@ -104,27 +102,4 @@ public:
 
 	_FORCE_INLINE_ int get_error_line() const { return err_line; }
 	_FORCE_INLINE_ String get_error_message() const { return err_str; }
-};
-
-class ResourceFormatLoaderJSON : public ResourceFormatLoader {
-	GDSOFTCLASS(ResourceFormatLoaderJSON, ResourceFormatLoader);
-
-public:
-	virtual Ref<Resource> load(const String &p_path, const String &p_original_path = "", Error *r_error = nullptr, bool p_use_sub_threads = false, float *r_progress = nullptr, CacheMode p_cache_mode = CACHE_MODE_REUSE) override;
-	virtual void get_recognized_extensions(List<String> *p_extensions) const override;
-	virtual bool handles_type(const String &p_type) const override;
-	virtual String get_resource_type(const String &p_path) const override;
-
-	// Treat JSON as a text file, do not generate a `*.json.uid` file.
-	virtual ResourceUID::ID get_resource_uid(const String &p_path) const override { return ResourceUID::INVALID_ID; }
-	virtual bool has_custom_uid_support() const override { return true; }
-};
-
-class ResourceFormatSaverJSON : public ResourceFormatSaver {
-	GDSOFTCLASS(ResourceFormatSaverJSON, ResourceFormatSaver);
-
-public:
-	virtual Error save(const Ref<Resource> &p_resource, const String &p_path, uint32_t p_flags = 0) override;
-	virtual void get_recognized_extensions(const Ref<Resource> &p_resource, List<String> *p_extensions) const override;
-	virtual bool recognize(const Ref<Resource> &p_resource) const override;
 };
