+++ core/register_core_types.cpp
@@ -35,10 +35,12 @@
 #include "core/core_bind.h"
 #include "core/crypto/aes_context.h"
 #include "core/crypto/crypto.h"
+#include "core/crypto/crypto_resource_format.h"
 #include "core/crypto/hashing_context.h"
 #include "core/debugger/engine_profiler.h"
 #include "core/extension/gdextension.h"
 #include "core/extension/gdextension_manager.h"
+#include "core/extension/gdextension_resource_format.h"
 #include "core/extension/godot_instance.h"
 #include "core/input/input.h"
 #include "core/input/input_map.h"
@@ -49,7 +51,9 @@
 #include "core/io/file_access_encrypted.h"
 #include "core/io/http_client.h"
 #include "core/io/image_loader.h"
+#include "core/io/image_resource_format.h"
 #include "core/io/json.h"
+#include "core/io/json_resource_format.h"
 #include "core/io/marshalls.h"
 #include "core/io/missing_resource.h"
 #include "core/io/packet_peer.h"
