+++ editor/scene/2d/tiles/tiles_editor_plugin.h
@@ -30,6 +30,7 @@
 
 #pragma once
 
+#include "core/os/semaphore.h"
 #include "editor/plugins/editor_plugin.h"
 #include "editor/scene/2d/tiles/tile_map_layer_editor.h"
 #include "editor/scene/2d/tiles/tile_set_editor.h"
