+++ modules/gltf/tests/test_gltf.h
@@ -44,6 +44,7 @@
 #include "editor/import/3d/resource_importer_scene.h"
 #include "editor/import/resource_importer_texture.h"
 #include "scene/resources/compressed_texture.h"
+#include "scene/resources/compressed_texture_resource_format.h"
 #include "scene/resources/packed_scene.h"
 #include "tests/test_utils.h"
