+++ modules/navigation_2d/nav_region_2d.h
@@ -34,6 +34,7 @@
 #include "nav_base_2d.h"
 #include "nav_utils_2d.h"
 
+#include "core/object/worker_thread_pool.h"
 #include "core/os/rw_lock.h"
 #include "scene/resources/2d/navigation_polygon.h"
